Affichage du filtre dans la page admin
